Remove plugin dependencies from theme
The theme should be able to function without any errors even if certain
plugins are not in place in the WordPress instance. This update checks
for the exists of certain actions and functions before attempting to
use them. This ensures that we can develop the theme in isolation
without having a bunch of notices and exceptions.